requirements.txt
pandas
networkx
openai
langchain
faiss-cpu
python-dotenv
fuzzywuzzy
